Update example to have a 'sprite' class
The example class has no sprite-like behaviour, but we are now working
in a framework where methods on an instance are called as event
handlers.

Each previous function is turned into a method; output lines are
prefixed with an instance attribute to check that an instance is
indeed being used; the class is registered as a 'sprite'.