copy.deepcopy(): Fix two bugs
Firstly, the 'return y' was mis-indented, such that an object with a
'__setstate__' would be copied as None.  Re-indent correctly.

Secondly, the duality between the JavaScript "$d" property and the
Python __dict__ attribute means that an object's __dict__ is
disconnected from its $d as a result of (simplifying)

    obj.__dict__.update(other.__dict__)

Fudge this by skipping the '__dict__' entry within an object's
__dict__.

TODO: This seems like a papering-over-the-cracks fix.  The appearance
of a '__dict__' entry in an object's __dict__ seems like it might
cause other problems elsewhere, too.  Log an issue/PR against upstream
Skulpt?

E.g.:

    class A:
        def __init__(self):
            self.banana = 42

    for k, v in A().__dict__.items():
        print(k, '=>', v)

prints

    banana => 42

in CPython(3) but

    __dict__ => {'__dict__': {...}, 'banana': 42}
    banana => 42

in Skulpt.

More directly incorrect, the code

    import copy

    class Alien:
        def __init__(self):
            self.id = 42

        def bump(self):
            self.id += 1

    a = Alien()
    print('a.id =', a.id)

    print('copying...')

    a1 = copy.deepcopy(a)
    a1.bump()
    print('a.id =', a.id)
    print('a1.id =', a1.id)
    print("a1.__dict__['id'] =", a1.__dict__['id'])

produces

    a.id = 42
    copying...
    a.id = 42
    a1.id = 43
    a1.__dict__['id'] = 43

as expected in CPython3, but

    a.id = 42
    copying...
    a.id = 42
    a1.id = 43
    a1.__dict__['id'] = 42

in Skulpt.